BUG: Fix bar sorting for ungrouped Vertical Bar charts
`git cherry-pick a00f237435344a77ee6e3a7c847ff65fda9a647d`
from master to AddeparCurrentVersion

When you provide data for a Vertical Bar chart, you can optionally provide
a `group` label for each data point, to indicate which bar group that bar
should be displayed as part of. This is as documented for the "data"
property of the vertical-bar component:
http://opensource.addepar.com/ember-charts/#/documentation

When all the data points have a null `group` or the same `group`,
we can say the chart is "ungrouped". In this case, we support
sorting the bars by their net value, by setting `sortKey: 'value'`.

Or, rather, we _claimed_ to support this. Unfortunately, `sortKey`
is effectively ignored in this case. The root cause is that we use
the same code to support ungrouped charts and grouped ones
(for which we do not support bar or group sorting). Hence, we always
laid out the bars in an ungrouped chart in alphabetical order by bar label,
the same way we lay out bars within a group of a grouped chart.

This changelist fixes this bug by differentiating between ungrouped
and grouped charts when computing an ordered list of bar labels,
whose order is then used as the order of the bars themselves.

**Compatibility WARNING:** If you used bar data points where _both_
the `group` label and the bar `label` were null or undefined,
these will no longer work with this change. I think it should
be easy to lift this restriction, however, so if it bothers you,
please let us know by filing an issue.

Testing: Passes Travis-CI gate including new test case:
    https://travis-ci.org/Addepar/ember-charts/builds/158570470
Code Reviewer: @billylittlefield
Closes #183